requirements.txt
Django==3.1.7
django-grappelli==2.15.1
djangorestframework==3.11.0
django-filter==2.4.0
GitPython==3.1.24